# core-common-customer-mgmt-sdk/src/main/resources/api-spec/openapi.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "core-common-customer-mgmt",
    "description": "Customer Management Core Application",
    "contact": {
      "name": "Firefly Software Foundation",
      "email": "dev@getfirefly.io"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://core.getfirefly.io/customers",
      "description": "Development Environment"
    },
    {
      "url": "/",
      "description": "Local Development Environment"
    }
  ],
...
  "paths": {
...
    "/api/v1/parties/{partyId}/addresses/{addressId}": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "Get address by ID",
        "description": "Retrieve a specific address associated with a party by its unique identifier",
        "operationId": "getAddressById",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "addressId",
            "in": "path",
            "description": "Unique identifier of the address",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Address or party not found"
          },
          "200": {
            "description": "Address found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AddressDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Addresses"
        ],
        "summary": "Update address",
        "description": "Update an existing address associated with a party",
        "operationId": "updateAddress",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "addressId",
            "in": "path",
            "description": "Unique identifier of the address",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid address data"
          },
          "404": {
            "description": "Address or party not found"
          },
          "200": {
            "description": "Address successfully updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AddressDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Addresses"
        ],
        "summary": "Delete address",
        "description": "Delete an address associated with a party from the system",
        "operationId": "deleteAddress",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "addressId",
            "in": "path",
            "description": "Unique identifier of the address",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Address or party not found"
          },
          "204": {
            "description": "Address successfully deleted"
          }
        }
      }
    },
...
    "/api/v1/parties/{partyId}/addresses": {
      "post": {
        "tags": [
          "Addresses"
        ],
        "summary": "Create address for a party",
        "description": "Create a new address associated with a specific party",
        "operationId": "createAddress",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Address successfully created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AddressDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid address data"
          },
          "404": {
            "description": "Party not found"
          }
        }
      }
    },
    "/api/v1/parties/{partyId}/addresses/filter": {
      "post": {
        "tags": [
          "Addresses"
        ],
        "summary": "Filter addresses for a party",
        "description": "Retrieve a paginated list of addresses associated with a specific party based on filtering criteria",
        "operationId": "filterAddresses",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "description": "Unique identifier of the party",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterRequestAddressDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Party not found"
          },
          "200": {
            "description": "Successfully retrieved filtered addresses",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter request"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddressDTO": {
        "required": [
          "addressKind",
          "city",
          "countryId",
          "line1",
          "partyId"
        ],
        "type": "object",
        "properties": {
          "addressId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "partyId": {
            "type": "string",
            "format": "uuid"
          },
          "addressKind": {
            "type": "string",
            "enum": [
              "HOME",
              "WORK",
              "MAILING",
              "REGISTERED"
            ]
          },
          "line1": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "line2": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "city": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "region": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "postalCode": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "countryId": {
            "type": "string",
            "format": "uuid"
          },
          "isPrimary": {
            "type": "boolean"
          },
          "latitude": {
            "maximum": 90,
            "exclusiveMaximum": false,
            "minimum": -90,
            "exclusiveMinimum": false,
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "maximum": 180,
            "exclusiveMaximum": false,
            "minimum": -180,
            "exclusiveMinimum": false,
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "FilterOptions": {
        "type": "object",
        "properties": {
          "caseInsensitiveStrings": {
            "type": "boolean"
          },
          "includeInheritedFields": {
            "type": "boolean"
          }
        },
        "description": "Filter options for customizing filter behavior"
      },
...
      "PaginationRequest": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "description": "The zero-based page number to retrieve.",
            "format": "int32",
            "example": 0,
            "default": 0
          },
          "pageSize": {
            "type": "integer",
            "description": "The number of items per page.",
            "format": "int32",
            "example": 10,
            "default": 10
          },
          "sortBy": {
            "type": "string",
            "description": "The field to sort the results by.",
            "example": "name"
          },
          "sortDirection": {
            "type": "string",
            "description": "The direction of sorting, either ASC or DESC.",
            "example": "DESC",
            "default": "DESC"
          }
        },
        "description": "Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction."
      },
      "RangeFilter": {
        "type": "object",
        "properties": {
          "ranges": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/RangeObject"
            },
            "description": "Map of field names to their range criteria"
          }
        },
        "description": "Represents range filters for numeric fields"
      },
      "RangeObject": {
        "type": "object",
        "properties": {
          "from": {
            "type": "object",
            "description": "Start value of the range"
          },
          "to": {
            "type": "object",
            "description": "End value of the range"
          }
        },
        "description": "Represents a range with from and to values"
      },
      "PaginationResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "description": "The list of items for the current page.",
            "items": {
              "type": "object",
              "description": "The list of items for the current page."
            }
          },
          "totalElements": {
            "type": "integer",
            "description": "The total number of elements across all pages.",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "description": "The total number of pages based on the data size and page size.",
            "format": "int32"
          },
          "currentPage": {
            "type": "integer",
            "description": "The current page number, typically zero-based.",
            "format": "int32"
          }
        },
        "description": "Represents a paginated response containing a list of items and pagination metadata."
      },
...
      "FilterRequestAddressDTO": {
        "required": [
          "pagination"
        ],
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/AddressDTO"
          },
          "rangeFilters": {
            "$ref": "#/components/schemas/RangeFilter"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationRequest"
          },
          "options": {
            "$ref": "#/components/schemas/FilterOptions"
          }
        },
        "description": "Filter criteria for addresses"
      },
...
    }
  }
}